refactor: rewrite js to morden style
