+++ themes/ssarcandy/js/ProjectPage.js
@@ -1,0 +1,2 @@
+import './Helper';
+import './ProjectTags';
