+++ themes/ssarcandy/layout/_partial/search.ejs
@@ -2,16 +2,4 @@
 <div class="search-panel" id="search-panel">
     <ul class="search-result" id="search-result"></ul>
 </div>
-<script type="text/template" id="search-tpl">
-<li class="item">
-    <a href="/{path}" class="waves-block waves-effect">
-        <div class="title ellipsis" title="{title}">{title}</div>
-        <div class="flex-row flex-middle">
-            <div class="tags ellipsis">
-                {tags}
-            </div>
-            <time class="flex-col time" style="text-align: right">{date}</time>
-        </div>
-    </a>
-</li>
-</script>
+
