+++ themes/ssarcandy/layout/layout.ejs
@@ -26,7 +26,7 @@
       <%- body %>
     </div>
   </main>
-<%- partial('_partial/after-footer') %>
 <%- partial('_partial/script') %>
+<%- partial('_partial/after-footer') %>
 </body>
 </html>
