+++ themes/ssarcandy/layout/works.ejs
@@ -14,3 +14,5 @@
 <script>
 window.SEARCH_INDEX = <%- lunr_index(site.data.works) %>;
 </script>
+
+<%- js('js/projectPage.bundle.js') %>
