+++ webpack.config.js
@@ -10,22 +10,23 @@ const isProd = NODE_ENV === 'production';
 
 module.exports = {
   mode: NODE_ENV,
-//   devtool: isProd ? 'source-map' : 'cheap-module-source-map',
+  devtool: isProd ? 'source-map' : 'cheap-module-source-map',
   entry: {
-    app: path.join(THEME_DIR, 'js/app.js')
+    app: path.join(THEME_DIR, 'js/App.js'),
+    projectPage: path.join(THEME_DIR, 'js/ProjectPage.js'),
   },
   output: {
-    filename: 'js/bundle.js',
+    filename: 'js/[name].bundle.js',
     path: path.join(THEME_DIR, 'source'),
     publicPath: '/'
   },
-//   optimization: {
-//     runtimeChunk: 'single',
-//     moduleIds: 'hashed',
-//     splitChunks: {
-//       chunks: 'all'
-//     }
-//   },
+  // optimization: {
+  //   runtimeChunk: 'single',
+  //   moduleIds: 'hashed',
+  //   splitChunks: {
+  //     chunks: 'all'
+  //   }
+  // },
   module: {
     rules: [
       {
